Fixed issue with days that do not appear in every month
Days of the month such as 30th of Feb and leap years are now handled
correctly. The library will now tell you leap years by setting the 29th
day of the month and Feb as filters. I've added a hard limit of 100
years to find a suitable time to run the cron because you can now create
crons that will never fire.